A few tweaks
Making the "Intro" Sec into "Among".  00 becomes "By".

Drill down on Event.Financing.... to give meaningful names to the choice between a Cap and no Cap (".Yes.sec" and ".No.sec").

Fix link for defined term "Equity Securities" (which is irregular in that the definition is plural, not the standard signular. So link is part of the Form, and not part of standard US/Def/Link/0.md.    (Fix is to remove an erronious carriage return.)